Refactor code and add examples to notebook in demonstrate_classes.py.
