+++ sock_demo.py
@@ -1,0 +1,8 @@
+# Demonstrate Python Knitting Pattern Calculator for a Sock
+import src.conversions as c
+import src.pattern as p
+import src.sock as s
+
+c.main()
+p.main()
+s.main()
